pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcm-engine"
version = "0.1.0"
description = "Memory Context Management engine — persistent knowledge management for AI coding sessions"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "unxmaal" },
]
keywords = ["mcp", "knowledge-management", "ai", "context", "session"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
mcm-engine = "mcm_engine.cli:main"

[project.entry-points."mcm_engine.plugins"]
# Example: my_plugin = "my_package:MyPlugin"

[tool.hatch.build.targets.wheel]
packages = ["src/mcm_engine"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
